# swagger.yaml
openapi: 3.0.3
info:
  title: Loto Key System API
  description: |
    ระบบหวยออนไลน์ - Hono Backend API

    **Authentication:** JWT Cookie (httpOnly) — ส่ง `credentials: 'include'` ทุก request

    **Response Format:**
    - Success: `{ success: true, data: {...} }`
    - Error: `{ success: false, message: "error message" }`
  version: 1.0.0
  contact:
    name: Loto Key System
servers:
  - url: http://localhost:3001/api
    description: Local Development Server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT token ใน httpOnly cookie
  schemas:
...
    Settings:
      type: object
      properties:
        isLotteryOpen:
          type: boolean
        closingTime:
          type: string
          example: "14:30"
        payRate2Digit:
          type: number
          example: 95
        payRate3Digit:
          type: number
          example: 900
        payRates:
          type: object
        autoReverse:
          type: boolean
        manualMode:
          type: boolean
        minBet:
          type: number
          example: 1
        maxBet2Digit:
          type: number
          example: 1000
        maxBet3Digit:
          type: number
          example: 500
...
    RestrictedNumber:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
          example: "69"
        maxAmount:
          type: number
          example: 500
        type:
          type: string
          example: "2ตัว"
        lotteryTypeId:
          type: integer
    RestrictedCreate:
      type: object
      required: [number, maxAmount]
      properties:
        number:
          type: string
        maxAmount:
          type: number
        type:
          type: string
        lotteryTypeId:
          type: integer
...
    LotteryType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "หวยไทย"
        code:
          type: string
          example: "thai"
        isOpen:
          type: boolean
        openTime:
          type: string
          example: "06:00"
        closeTime:
          type: string
          example: "14:30"
        icon:
          type: string
        sortOrder:
          type: integer
    LotteryTypeCreate:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        openTime:
          type: string
          example: "06:00"
        closeTime:
          type: string
          example: "14:30"
        icon:
          type: string
...
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: ดึงการตั้งค่าระบบ
      description: "ใช้ใน: keyPage, settingsPage, lotterySelectPage | Permission: settings.read"
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Settings data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Settings'

    put:
      tags: [Settings]
      summary: อัปเดตการตั้งค่า
      description: "ใช้ใน: settingsPage (toggle, rates, limits, time, reverse, manual) | Permission: settings.write"
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true

  # ============================================================
  # TRANSACTIONS
  # ============================================================
...
  /restricted:
    get:
      tags: [Restricted]
      summary: ดึงรายการหวยอั้น
      description: "ใช้ใน: settingsPage | Permission: numbers.read"
      security:
        - cookieAuth: []
      parameters:
        - name: lotteryTypeId
          in: query
          schema:
            type: integer
          description: กรองตามประเภทหวย
      responses:
        '200':
          description: List of restricted numbers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestrictedNumber'

    post:
      tags: [Restricted]
      summary: เพิ่มเลขอั้น
      description: "ใช้ใน: settingsPage | Permission: numbers.create"
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestrictedCreate'
      responses:
        '200':
          description: Added — returns updated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestrictedNumber'
  /restricted/check/{number}:
    get:
      tags: [Restricted]
      summary: ตรวจสอบเลขอั้น
      description: "ใช้ใน: keyPage | Permission: numbers.read"
      security:
        - cookieAuth: []
      parameters:
        - name: number
          in: path
          required: true
          schema:
            type: string
          description: เลขที่ต้องการตรวจ
        - name: lotteryTypeId
          in: query
          schema:
            type: integer
          description: ตรวจภายในประเภทหวย
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  restricted:
                    type: boolean
                  data:
                    type: object
                    properties:
                      maxAmount:
                        type: number
  /restricted/id/{id}:
    delete:
      tags: [Restricted]
      summary: ลบเลขอั้นตาม ID
      description: "ใช้ใน: settingsPage | Permission: numbers.delete"
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removed — returns updated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestrictedNumber'
  /restricted/{number}:
    delete:
      tags: [Restricted, Unused]
      summary: ลบเลขอั้นตามเลข (Legacy)
      description: "⚠️ ไม่ได้ใช้งานจาก Frontend — ถูกแทนที่ด้วย DELETE /restricted/id/:id | Permission: numbers.delete"
      security:
        - cookieAuth: []
      parameters:
        - name: number
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed

  # ============================================================
  # BET TYPES
  # ============================================================
...
  /lottery-types:
    get:
      tags: [LotteryTypes]
      summary: ดึงประเภทหวยทั้งหมด
      description: "ใช้ใน: settingsPage, lotterySelectPage"
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of lottery types sorted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LotteryType'

    post:
      tags: [LotteryTypes]
      summary: เพิ่มประเภทหวย
      description: "ใช้ใน: settingsPage"
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LotteryTypeCreate'
      responses:
        '201':
          description: Lottery type created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/LotteryType'
  /lottery-types/check/{code}:
    get:
      tags: [LotteryTypes]
      summary: ตรวจสอบประเภทหวย (เปิด/ปิด)
      description: "ใช้ใน: keyPage"
      security:
        - cookieAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          description: "รหัสประเภทหวย เช่น 'thai', 'laos'"
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      exists:
                        type: boolean
                      isOpen:
                        type: boolean
                      name:
                        type: string
                      closeTime:
                        type: string
  /lottery-types/{id}:
    patch:
      tags: [LotteryTypes]
      summary: อัปเดตประเภทหวย
      description: "ใช้ใน: settingsPage (toggle เปิด/ปิด, แก้เวลา)"
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                isOpen:
                  type: boolean
                openTime:
                  type: string
                closeTime:
                  type: string
      responses:
        '200':
          description: Lottery type updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/LotteryType'

    delete:
      tags: [LotteryTypes]
      summary: ลบประเภทหวย
      description: "ใช้ใน: settingsPage"
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Deleted"
  /lottery-types/{id}/logs:
    get:
      tags: [LotteryTypes]
      summary: ดึงประวัติการเปลี่ยนแปลง
      description: "ใช้ใน: settingsPage"
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Change logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        action:
                          type: string
                        changes:
                          type: object
                        createdAt:
                          type: string

  # ============================================================
  # ROUNDS
  # ============================================================
...